refactor(ingredient-measure): move to observer and move specs to correct file
