+++ app/models/ingredient_measure.rb
@@ -1,34 +1,9 @@
 class IngredientMeasure < ApplicationRecord
+  include PowerTypes::Observable
   belongs_to :ingredient
 
   validates :name, presence: true, uniqueness: { scope: :ingredient_id }
   validates :quantity, presence: true
-
-  before_save do
-    name.capitalize!
-    update_recipe_ingredient_if_exists
-  end
-
-  before_destroy { change_recipe_ingredient_if_exists }
-
-  def update_recipe_ingredient_if_exists
-    ingredient.recipe_ingredients.each do |recipe_ingredient|
-      next unless recipe_ingredient.ingredient_measure == name_was
-
-      recipe_ingredient.update!(ingredient_measure: name)
-    end
-  end
-
-  def change_recipe_ingredient_if_exists
-    ingredient.recipe_ingredients.each do |recipe_ingredient|
-      next unless recipe_ingredient.ingredient_measure == name_was
-
-      replacing_measure = ingredient.default_measure
-      name = replacing_measure.name
-      quantity = recipe_ingredient.ingredient_quantity / quantity_was * replacing_measure.quantity
-      recipe_ingredient.update!(ingredient_measure: name, ingredient_quantity: quantity)
-    end
-  end
 end
 
 # == Schema Information
