+++ spec/models/ingredient_spec.rb
@@ -37,38 +37,4 @@ RSpec.describe Ingredient, type: :model do
       expect(described_class.below_minimum.count).to eq(ingredients_bad.count)
     end
   end
-
-  describe 'destroy of measure does not break recipe logic' do
-    let!(:recipe) { create(:recipe) }
-    let!(:ingredient) { create(:ingredient) }
-    let!(:first_ingredient_measure) do
-      create(:ingredient_measure, ingredient: ingredient, name: 'Kilo', quantity: 3, primary: true)
-    end
-    let!(:second_ingredient_measure) do
-      create(:ingredient_measure, ingredient: ingredient, name: 'Gramo', quantity: 3000)
-    end
-    let!(:recipe_ingredient) do
-      create(:recipe_ingredient, ingredient: ingredient, recipe: recipe,
-                                 ingredient_measure: 'Gramo', ingredient_quantity: 1500)
-    end
-
-    before do
-      ingredient.reload
-      recipe.reload
-    end
-
-    it 'changes ingredient measure name that is on recipe, so changes the name also' do
-      second_ingredient_measure.update!(name: 'Gramito')
-
-      expect(recipe_ingredient.reload.ingredient_measure).to eq('Gramito')
-      expect(recipe_ingredient.reload.ingredient_quantity).to eq(1500)
-    end
-
-    it 'deletes ingredient measure that is on recipe, but it changes to the existing one' do
-      second_ingredient_measure.destroy!
-
-      expect(recipe_ingredient.reload.ingredient_measure).to eq('Kilo')
-      expect(recipe_ingredient.reload.ingredient_quantity).to eq(1.5)
-    end
-  end
 end
